+++ src/karboni/database/library.py
@@ -2,6 +2,7 @@ import datetime
 import logging
 from collections.abc import Callable
 from functools import wraps
+from itertools import chain
 from types import TracebackType
 from typing import Any, TypeVar
 
@@ -321,9 +322,25 @@ class Library:
             )
             self._session.merge(collection_object)
 
+    def _find_subcollection(self, collection_key: str) -> list[str]:
+        descendants = []
+        queue = [collection_key]
+        while queue:
+            current = queue.pop(0)
+            stmt = select(Collection.collection_key).where(Collection.parent_collection == current)
+            children = list(self._session.scalars(stmt).all())
+            descendants.extend(children)
+            queue.extend(children)
+        return descendants
+
     @write_operation
     def bulk_delete_collections(self, keys: list[str]) -> None:
-        self._bulk_delete_objects(Collection, "collection_key", keys)
+        # When a collection is deleted, Zotero doesn't keep track of its deleted subcollections,
+        # thus we must find them, if any, and add them to the deletion list.
+        subcollections = chain(*[self._find_subcollection(key) for key in keys])
+        self._bulk_delete_objects(
+            Collection, "collection_key", list(set(keys) | set(subcollections))
+        )
 
     @write_operation
     def bulk_insert_collections(self, collections: list[Any]) -> None: